Moved Config to separate file
