+++ trunk-recorder/global_structs.h
@@ -27,6 +27,7 @@ struct Config {
   std::string temp_dir;
   std::string debug_recorder_address;
   std::string log_dir;
+  std::string default_mode;
   bool new_call_from_update;
   bool debug_recorder;
   int debug_recorder_port;
